nginx-all-practice/nginx-python-pg/py/requirements.txt
flask
psycopg2
python-dotenv